requirements.txt
flash-attn
decord
tqdm>=4.66.1
absl-py==2.1.0
accelerate==1.1.1
asttokens==2.4.1
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.10.0
grpcio==1.62.2
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.9.3
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numexpr==2.10.1
numpy==1.26.4
nvidia-ml-py==12.535.133
nvitop==1.3.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
stack-data==0.6.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard_data_server==0.7.0
termcolor==2.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
zipp==3.21.0
